SQL fixes
- Consistent SQL for new installations and updates
- Larger data type for id and itemId columns
- No display width for integer columns on MySQL
- Use order number for schemaorg plugins
- Fix order numbers of system plugins on new installation
- Fix character set and collation for MySQL on new installation
- Align code style to existing SQL

+++ installation/sql/postgresql/extensions.sql
@@ -899,7 +899,7 @@ CREATE INDEX "#__scheduler_tasks_idx_checked_out" ON "#__scheduler_tasks" ("chec
 CREATE TABLE IF NOT EXISTS "#__schemaorg"
 (
   "id" serial NOT NULL,
-  "itemId" integer,
+  "itemId" bigint,
   "context" varchar(100),
   "schemaType" varchar(100),
   "schemaForm" text,
